# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lykke.day",
    "description": "API for managing calendar events",
    "version": "0.1.0"
  },
  "paths": {
...
    "/planning/routines/": {
      "get": {
        "tags": [
          "routine"
        ],
        "summary": "List Routines",
        "description": "Get all routines for the current user.",
        "operationId": "list_routines_planning_routines__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoutineSchema"
                  },
                  "type": "array",
                  "title": "Response List Routines Planning Routines  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/planning/preview/today": {
      "get": {
        "tags": [
          "routine"
        ],
        "summary": "Preview Today",
        "description": "Preview what today would look like if scheduled.",
        "operationId": "preview_today_planning_preview_today_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayContextSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/planning/tomorrow/preview": {
      "get": {
        "tags": [
          "routine"
        ],
        "summary": "Preview Tomorrow",
        "description": "Preview what tomorrow would look like if scheduled.",
        "operationId": "preview_tomorrow_planning_tomorrow_preview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayContextSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/planning/date/{date}/preview": {
      "get": {
        "tags": [
          "routine"
        ],
        "summary": "Preview Date",
        "description": "Preview what a specific date would look like if scheduled.",
        "operationId": "preview_date_planning_date__date__preview_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayContextSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/planning/schedule/today": {
      "put": {
        "tags": [
          "routine"
        ],
        "summary": "Schedule Today",
        "description": "Schedule today with tasks from routines.",
        "operationId": "schedule_today_planning_schedule_today_put",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayContextSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/planning/tomorrow/schedule": {
      "put": {
        "tags": [
          "routine"
        ],
        "summary": "Schedule Tomorrow",
        "description": "Schedule tomorrow with tasks from routines.",
        "operationId": "schedule_tomorrow_planning_tomorrow_schedule_put",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayContextSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/planning/date/{date}/schedule": {
      "put": {
        "tags": [
          "routine"
        ],
        "summary": "Schedule Date",
        "description": "Schedule a specific date with tasks from routines.",
        "operationId": "schedule_date_planning_date__date__schedule_put",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayContextSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/push/subscriptions/": {
      "get": {
        "tags": [
          "push",
          "notifications"
        ],
        "summary": "List Subscriptions",
        "operationId": "list_subscriptions_push_subscriptions__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PushSubscriptionSchema"
                  },
                  "type": "array",
                  "title": "Response List Subscriptions Push Subscriptions  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/push/subscriptions/{subscription_id}": {
      "delete": {
        "tags": [
          "push",
          "notifications"
        ],
        "summary": "Delete Subscription",
        "operationId": "delete_subscription_push_subscriptions__subscription_id__delete",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subscription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/push/subscribe/": {
      "post": {
        "tags": [
          "push",
          "notifications"
        ],
        "summary": "Subscribe",
        "operationId": "subscribe_push_subscribe__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscriptionSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
...
    "/routines/": {
      "get": {
        "tags": [
          "routines"
        ],
        "summary": "List Routines",
        "description": "List routines with pagination.",
        "operationId": "list_routines_routines__get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedQueryResponse_RoutineSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "routines"
        ],
        "summary": "Create Routine",
        "description": "Create a new routine.",
        "operationId": "create_routine_routines__post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineCreateSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalendarEntrySchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "calendar_id": {
            "type": "string",
            "format": "uuid",
            "title": "Calendar Id"
          },
          "platform_id": {
            "type": "string",
            "title": "Platform Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "frequency": {
            "$ref": "#/components/schemas/TaskFrequency"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionSchema"
            },
            "type": "array",
            "title": "Actions"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name",
          "calendar_id",
          "platform_id",
          "platform",
          "status",
          "starts_at",
          "frequency",
          "created_at",
          "updated_at",
          "date"
        ],
        "title": "CalendarEntrySchema",
        "description": "API schema for CalendarEntry entity."
      },
...
      "DayContextSchema": {
        "properties": {
          "day": {
            "$ref": "#/components/schemas/DaySchema"
          },
          "calendar_entries": {
            "items": {
              "$ref": "#/components/schemas/CalendarEntrySchema"
            },
            "type": "array",
            "title": "Calendar Entries"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSchema"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "day"
        ],
        "title": "DayContextSchema",
        "description": "API schema for DayContext value object."
      },
...
      "DaySchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "alarm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlarmSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/DayStatus"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/DayTag"
            },
            "type": "array",
            "title": "Tags"
          },
          "template": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DayTemplateSchema"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "date",
          "status"
        ],
        "title": "DaySchema",
        "description": "API schema for Day entity."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PagedQueryResponse_RoutineSchema_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RoutineSchema"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "has_next",
          "has_previous"
        ],
        "title": "PagedQueryResponse"
      },
      "PushSubscriptionSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "p256dh": {
            "type": "string",
            "title": "P256Dh"
          },
          "auth": {
            "type": "string",
            "title": "Auth"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "endpoint",
          "p256dh",
          "auth",
          "created_at"
        ],
        "title": "PushSubscriptionSchema",
        "description": "API schema for PushSubscription entity."
      },
      "RoutineCreateSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/TaskCategory"
          },
          "routine_schedule": {
            "$ref": "#/components/schemas/RoutineSchedule"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/RoutineTask-Input"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category",
          "routine_schedule"
        ],
        "title": "RoutineCreateSchema",
        "description": "API schema for creating a Routine entity."
      },
      "RoutineSchedule": {
        "properties": {
          "frequency": {
            "$ref": "#/components/schemas/TaskFrequency"
          },
          "weekdays": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DayOfWeek"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekdays"
          },
          "day_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Number"
          }
        },
        "type": "object",
        "required": [
          "frequency"
        ],
        "title": "RoutineSchedule"
      },
      "RoutineSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/TaskCategory"
          },
          "routine_schedule": {
            "$ref": "#/components/schemas/RoutineSchedule"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/RoutineTask-Output"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name",
          "category",
          "routine_schedule"
        ],
        "title": "RoutineSchema",
        "description": "API schema for Routine entity."
      },
      "RoutineTask-Input": {
        "properties": {
          "task_definition_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Definition Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskSchedule"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "task_definition_id"
        ],
        "title": "RoutineTask"
      },
      "RoutineTask-Output": {
        "properties": {
          "task_definition_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Definition Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskSchedule"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "task_definition_id"
        ],
        "title": "RoutineTask"
      },
...
      "SubscriptionRequest": {
        "properties": {},
        "type": "object",
        "title": "SubscriptionRequest"
      },
      "TaskCategory": {
        "type": "string",
        "enum": [
          "HYGIENE",
          "GROOMING",
          "SLEEP",
          "HEALTH",
          "FITNESS",
          "MEDICAL",
          "NUTRITION",
          "COOKING",
          "HOUSE",
          "CLEANING",
          "MAINTENANCE",
          "ORGANIZATION",
          "WORK",
          "MEETING",
          "PROFESSIONAL",
          "FAMILY",
          "SOCIAL",
          "RELATIONSHIP",
          "SHOPPING",
          "ERRAND",
          "COMMUTE",
          "TRAVEL",
          "ENTERTAINMENT",
          "HOBBY",
          "RECREATION",
          "EDUCATION",
          "LEARNING",
          "FINANCE",
          "BILLS",
          "TECHNOLOGY",
          "DIGITAL",
          "PET",
          "PLANNING",
          "ADMIN"
        ],
        "title": "TaskCategory"
      },
...
      "TaskFrequency": {
        "type": "string",
        "enum": [
          "DAILY",
          "CUSTOM_WEEKLY",
          "WEEKLY",
          "ONCE",
          "YEARLY",
          "MONTHLY",
          "BI_WEEKLY",
          "WORK_DAYS",
          "WEEKENDS"
        ],
        "title": "TaskFrequency"
      },
...
      "TaskSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "scheduled_date": {
            "type": "string",
            "format": "date",
            "title": "Scheduled Date"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "task_definition": {
            "$ref": "#/components/schemas/TaskDefinitionSchema"
          },
          "category": {
            "$ref": "#/components/schemas/TaskCategory"
          },
          "frequency": {
            "$ref": "#/components/schemas/TaskFrequency"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskScheduleSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "routine_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routine Id"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TaskTag"
            },
            "type": "array",
            "title": "Tags"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionSchema"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "scheduled_date",
          "name",
          "status",
          "task_definition",
          "category",
          "frequency"
        ],
        "title": "TaskSchema",
        "description": "API schema for Task entity."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lykke_auth"
      }
    }
  }
}